pyproject.toml
[project]
name = "cms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyqt6-tools>=6.4.2.3.3",
]
